dynaml-tensorflow: Refactor to `dynaml.tensorflow.utils` package; created `dtfutils` object.
