+++ dynaml-core/src/main/scala-2.11/io/github/mandar2812/dynaml/evaluation/RegressionMetricsTF.scala
@@ -20,7 +20,7 @@ package io.github.mandar2812.dynaml.evaluation
 
 import com.quantifind.charts.Highcharts.{regression, title, xAxis, yAxis}
 import io.github.mandar2812.dynaml.tensorflow.dtf
-import io.github.mandar2812.dynaml.tensorflow.{utils => dtfutils}
+import io.github.mandar2812.dynaml.tensorflow.utils.dtfutils
 import org.platanios.tensorflow.api.{::, Tensor}
 
 /**
